pyproject.toml
[tool.poetry]
name = "falca"
version = "0.1.0"
description = "wip"
authors = ["aprilahijriyan <hijriyan23@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
falcon = "^3.0.0"
apispec = "^4.4.0"
webargs = "^8.0.0"
Mako = "^1.1.4"
six = "^1.15.0"
orjson = {version = "^3.5.2", extras = ["full"]}
python-rapidjson = {version = "^1.0", extras = ["full"]}

[tool.poetry.dev-dependencies]
ipython = "^7.22.0"
gunicorn = "^20.1.0"
uvicorn = "^0.13.4"
pre-commit = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
